standardized output file names, created standalone mintaka-wikidata folder, moved parseCSVFile to own file
