+++ src/utils/evaluations/mintaka-wikidata/data/.gitignore
@@ -1,0 +1,1 @@
+*.csv
