--- src/utils/evaluations/questions.ts
@@ -1,6 +1,0 @@
-// Copyright (c) 2024 Massachusetts Institute of Technology
-// SPDX-License-Identifier: MIT
-
-import { MintakaQuestionType } from "./mintakaEvaluation";
-
-export const QUESTIONS:MintakaQuestionType[] = []
